Fix compilation warning
Fix warning about comparison between signed and unsigned integer in
proto.c by changing idx from int to unsigned int.

+++ proto.c
@@ -27,7 +27,7 @@
 
 /* do we really need all these?? */
 
-static int idx;
+static size_t idx;
 static const unsigned char protos[] = {
 	"\000ip\0"
 	"\001icmp\0"
